# Order-service/src/main/java/org/orderservice/model/OrderItem.java
package org.orderservice.model;

import lombok.*;
import org.springframework.data.annotation.Id;
import org.springframework.data.relational.core.mapping.Table;

import java.math.BigDecimal;

@Table(name = "order_items")
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class OrderItem {
    @Id
    private int id;
    private int orderId;
    private String productSku;
    private String productName;
    private int quantity;
    private BigDecimal price;
}